requirements.txt
langchain
openai
python-dotenv
pinecone-client
tiktoken
pypdf
streamlit-extras
unstructured
Flask